--- src/components/button/Button/index.type.ts
@@ -1,12 +1,0 @@
-import { FontSizeType } from '@components/text/index.type';
-import { ColorType } from '@utils/color';
-
-export interface ButtonProps {
-  children?: React.ReactNode;
-  labelColor?: ColorType;
-  color?: ColorType;
-  type?: 'default' | 'round';
-  size?: FontSizeType;
-  width?: 'default' | 'full';
-  onClick?: () => void;
-}
